requirements.txt
Django==4.2.16
mysqlclient==2.2.4
python-decouple==3.8
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
lifelines==0.29.0
xgboost==2.1.3
optuna==3.6.1
shap==0.45.1
scikit-learn==1.5.2
sdv==1.14.0
pandas==2.2.2
numpy==1.26.4
scipy==1.13.1
joblib==1.4.2
reportlab==4.2.2
pdfplumber==0.11.4
Pillow==10.4.0
whitenoise==6.7.0
gunicorn==22.0.0
matplotlib==3.9.2
